Fixed Formatting Errors
An HTML div element was interfering with the rendering of a portion of the Markdown. Also, the Latex code is now an image.
Note: I do not know if it is better to save the image to the server or just fetch it from the third-party each time the page is loaded. I would prefer the latter, but am not in a position to do that right now.